bullets will now properly bounce off of wall verts
when a bullet bounces off of a wall, moving outside of the collision with the new angle now only moves slightly farther than the ball's current sub-step speed. If the ball has not escaped its collision in that time, we assume that the collision must have, in fact, been with the thin edge of the wall, so we reset and try the collision again with the wall angle offset by 90 degrees (pi/2 radians).